day layout algorithm as strategy pattern
