+++ examples/events.js
@@ -111,4 +111,52 @@ export default [
     start: now,
     end: now,
   },
+  {
+    id: 16,
+    title: 'Video Record',
+    start: new Date(2015, 3, 14, 15, 30, 0),
+    end: new Date(2015, 3, 14, 19, 0, 0),
+  },
+  {
+    id: 17,
+    title: 'Dutch Song Producing',
+    start: new Date(2015, 3, 14, 16, 30, 0),
+    end: new Date(2015, 3, 14, 20, 0, 0),
+  },
+  {
+    id: 18,
+    title: 'Itaewon Halloween Meeting',
+    start: new Date(2015, 3, 14, 16, 30, 0),
+    end: new Date(2015, 3, 14, 17, 30, 0),
+  },
+  {
+    id: 19,
+    title: 'Online Coding Test',
+    start: new Date(2015, 3, 14, 17, 30, 0),
+    end: new Date(2015, 3, 14, 20, 30, 0),
+  },
+  {
+    id: 20,
+    title: 'An overlapped Event',
+    start: new Date(2015, 3, 14, 17, 0, 0),
+    end: new Date(2015, 3, 14, 18, 30, 0),
+  },
+  {
+    id: 21,
+    title: 'Phone Interview',
+    start: new Date(2015, 3, 14, 17, 0, 0),
+    end: new Date(2015, 3, 14, 18, 30, 0),
+  },
+  {
+    id: 22,
+    title: 'Cooking Class',
+    start: new Date(2015, 3, 14, 17, 30, 0),
+    end: new Date(2015, 3, 14, 19, 0, 0),
+  },
+  {
+    id: 23,
+    title: 'Go to the gym',
+    start: new Date(2015, 3, 14, 18, 30, 0),
+    end: new Date(2015, 3, 14, 20, 0, 0),
+  },
 ]
