+++ src/Calendar.js
@@ -6,6 +6,7 @@ import {
   accessor,
   dateFormat,
   dateRangeFormat,
+  DayLayoutAlgorithmPropType,
   views as componentViews,
 } from './utils/propTypes'
 import warning from 'warning'
@@ -718,6 +719,14 @@ class Calendar extends React.Component {
       noEventsInRange: PropTypes.node,
       showMore: PropTypes.func,
     }),
+
+    /**
+     * A day event layout(arrangement) algorithm.
+     * `overlap` allows events to be overlapped.
+     * `no-overlap` resizes events to avoid overlap.
+     * or custom `Function(events, minimumStartDifference, slotMetrics, accessors)`
+     */
+    dayLayoutAlgorithm: DayLayoutAlgorithmPropType,
   }
 
   static defaultProps = {
@@ -743,6 +752,7 @@ class Calendar extends React.Component {
 
     longPressThreshold: 250,
     getNow: () => new Date(),
+    dayLayoutAlgorithm: 'overlap',
   }
 
   constructor(...args) {
