+++ src/utils/propTypes.js
@@ -39,3 +39,8 @@ export let views = PropTypes.oneOfType([
     }
   }),
 ])
+
+export const DayLayoutAlgorithmPropType = PropTypes.oneOfType([
+  PropTypes.oneOf(['overlap', 'no-overlap']),
+  PropTypes.func,
+])
